+++ tests/v1/e2e/general/test_cascade_attention.py
@@ -5,7 +5,7 @@ import pytest
 
 from vllm import LLM, SamplingParams
 
-from ...utils import create_new_process_for_each_test
+from ....utils import create_new_process_for_each_test
 
 
 @create_new_process_for_each_test()
